Changed db from h2 to postgres, deleted data.sql file, added getClan feature, changed id type from Long to Integer

+++ src/main/java/clangame/model/Clan.java
@@ -7,9 +7,8 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-
 public class Clan {
-    private Long id;
+    private Integer id;
     private String name;
     private int gold;
 }
